+++ docs/.vitepress/config/sidebar.ts
@@ -105,7 +105,6 @@ export const sidebar = {
         { text: 'Dialog 对话框', link: '/components/dialog' },
         { text: 'Drawer 抽屉', link: '/components/drawer' },
         { text: 'Tooltip 消息提示', link: '/components/tooltip' },
-        { text: 'Popup 弹出窗口', link: '/components/popup' },
         { text: 'Swap 切换', link: '/components/swap' }
       ]
     },