+++ packages/fighting-design/drawer/src/drawer.vue
@@ -1,9 +1,9 @@
 <script lang="ts" setup>
   import { Props } from './props'
-  import { toRef } from 'vue'
-  import { useVisible } from '../../_hooks'
   import { FCloseBtn } from '../../close-btn'
-  import { FPopup } from '../../popup'
+  import { useRun } from '../../_hooks'
+  import { toRef } from 'vue'
+  import { useVisible, useList } from '../../_hooks'
   import { EMIT_VISIBLE } from '../../_tokens'
 
   defineOptions({ name: 'FDrawer' })
@@ -11,44 +11,113 @@
   const prop = defineProps(Props)
   const emit = defineEmits([EMIT_VISIBLE])
 
+  const { run } = useRun()
+  const { styles, classes } = useList(prop, 'drawer')
   const { isVisible, closeVisible } = useVisible(toRef(prop, 'visible'), emit)
+
+  /** 样式列表 */
+  const styleList = styles(['width'])
+
+  /** 类名列表 */
+  const classList = classes(['fullscreen', 'maskBlur', 'direction'], 'f-drawer')
+
+  /**
+   * 打开动画开始执行的回调
+   *
+   * @param { Object } node 元素节点
+   */
+  const handleOpen = (node: Element): void => {
+    /** 开启状态隐藏滚动条 */
+    document.body.style.overflow = 'hidden'
+    run(prop.onOpen, node)
+  }
+
+  /**
+   * 打开动画结束执行的回调
+   *
+   * @param { Object } node 元素节点
+   */
+  const handleOpenEnd = (node: Element): void => {
+    run(prop.onOpenEnd, node)
+  }
+
+  /**
+   * 关闭动画开始执行的回调
+   *
+   * @param { Object } node 元素节点
+   */
+  const handleClose = (node: Element): void => {
+    document.body.style.overflow = ''
+    run(prop.onClose, node)
+  }
+
+  /**
+   * 关闭动画结束执行的回调
+   *
+   * @param { Object } node 元素节点
+   */
+  const handleCloseEnd = (node: Element): void => {
+    run(prop.onCloseEnd, node)
+  }
+
+  /** 点击遮罩层关闭 */
+  const closePopup = (): void => {
+    if (!prop.maskClose) return
+    emit(EMIT_VISIBLE, false)
+  }
 </script>
 
 <template>
-  <f-popup
-    v-model:visible="isVisible"
-    :append-to-body="appendToBody"
-    :show-mask="showMask"
-    :mask-close="maskClose"
-    :z-index="zIndex"
-    :mask-blur="maskBlur"
-    :direction="direction"
-    :popup-size="size"
-    :on-open="onOpen"
-    :on-open-end="onOpenEnd"
-    :on-close="onClose"
-    :on-close-end="onCloseEnd"
-  >
-    <div class="f-drawer">
-      <!-- 头部 -->
-      <header class="f-drawer__header">
-        <slot name="header">
-          <span class="f-drawer__header-title">{{ title }}</span>
-          <f-close-btn v-if="showCloseIcon" :icon="closeIcon" :on-click="closeVisible">
-            <slot name="closeIcon" />
-          </f-close-btn>
-        </slot>
-      </header>
+  <teleport to="body" :disabled="!appendToBody">
+    <transition
+      name="f-drawer__trans"
+      @before-enter="handleOpen"
+      @after-enter="handleOpenEnd"
+      @before-leave="handleClose"
+      @after-leave="handleCloseEnd"
+    >
+      <div
+        v-show="isVisible"
+        role="dialog"
+        aria-modal="true"
+        tabindex="-1"
+        :class="classList"
+        :style="styleList"
+      >
+        <!-- 遮罩层 -->
+        <div v-if="showMask" class="f-drawer__mask" />
 
-      <!-- 身体 -->
-      <section v-if="$slots.default" class="f-drawer__body">
-        <slot />
-      </section>
+        <!-- 主内容 -->
+        <div class="f-drawer__wrapper" @click.self="closePopup">
+          <transition name="f-drawer__container-trans">
+            <div v-show="isVisible" class="f-drawer__container">
+              <!-- 头部 -->
+              <header class="f-drawer__header">
+                <slot name="header">
+                  <span class="f-drawer__header-title">{{ title }}</span>
+                  <f-close-btn
+                    v-if="showCloseIcon"
+                    :icon="closeIcon"
+                    :on-click="closeVisible"
+                  >
+                    <slot name="closeIcon" />
+                  </f-close-btn>
+                </slot>
+              </header>
 
-      <!-- 页脚 -->
-      <footer v-if="$slots.footer" class="f-drawer__footer">
-        <slot name="footer" />
-      </footer>
-    </div>
-  </f-popup>
+              <!-- 身体 -->
+              <section v-if="$slots.default" class="f-drawer__body">
+                <slot />
+              </section>
+
+              <!-- 页脚 -->
+              <footer v-if="$slots.footer" class="f-drawer__footer">
+                <slot name="footer" />
+              </footer>
+            </div>
+          </transition>
+        </div>
+      </div>
+    </transition>
+  </teleport>
 </template>